requirements.txt
ollama
langchain-core
langchain-ollama
jupyter
numpy>=1.26.0,<2.0.0
langchain==0.3.1
langchain-community==0.3.1
llama-index-core==0.11.15
fastapi[all]
matplotlib
pandas
sse-starlette
gunicorn